Refactor Clairvoyance
