+++ app/agents/voice/automatic/features/charts/rtvi/rtvi.py
@@ -1,7 +1,7 @@
 from pipecat.processors.frameworks.rtvi import RTVIProcessor, RTVIServerMessageFrame
 
 from app.core.logger import logger
-from app.agents.voice.automatic.services.charts.chart_tools import get_pending_chart_emissions
+from app.agents.voice.automatic.features.charts.chart_tools import get_pending_chart_emissions
 
 async def emit_chart_components(rtvi: RTVIProcessor, function_name: str, session_id: str) -> None:
     """Emit chart components via RTVI frames after function calls."""
