+++ app/agents/voice/automatic/features/llm_wrapper.py
@@ -1,10 +1,10 @@
 import uuid
 from typing import List, Dict, Any, Optional, Callable
-from app.agents.voice.automatic.services.context_summarizer import ContextSummarizer
+from app.agents.voice.automatic.features.summarizer.context_summarizer import ContextSummarizer
 from app.core import config
 from app.core.logger import logger
-from app.agents.voice.automatic.services.hitl.utils import is_dangerous_operation
-from app.agents.voice.automatic.services.hitl.hitl import get_hitl_manager
+from app.agents.voice.automatic.features.hitl.utils import is_dangerous_operation
+from app.agents.voice.automatic.features.hitl.hitl import get_hitl_manager
 from app.core.config import HITL_ENABLE
 class LLMServiceWrapper:
     def __init__(self, llm_service):
