+++ app/agents/voice/automatic/utils/conversation_manager.py
@@ -6,12 +6,10 @@ Handles session-based conversation storage, memory management, and cleanup.
 import asyncio
 import time
 from typing import Dict, List, Optional, Any
-from collections import defaultdict
 import threading
-from datetime import datetime, timedelta
 
 from app.core.logger import logger
-from app.agents.voice.automatic.services.charts.conversation import (
+from app.agents.voice.automatic.features.charts.conversation import (
     ConversationDebugData, 
     ConversationTurn, 
     ConversationMessage,
